Exceptions & test services

+++ src/main/java/edu/eci/ieti/triddy/services/impl/UserServiceImpl.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import edu.eci.ieti.triddy.exceptions.TriddyServiceException;
+import edu.eci.ieti.triddy.exceptions.UserNotFoundException;
 import edu.eci.ieti.triddy.model.User;
 import edu.eci.ieti.triddy.repository.UserRepository;
 import edu.eci.ieti.triddy.services.UserService;
@@ -26,52 +28,82 @@ public class UserServiceImpl implements UserService{
     }
 
     @Override
-    public User getUser(String email) {
-        return userRepository.findByEmail(email);
+    public User getUser(String email) throws UserNotFoundException {
+        User res = userRepository.findByEmail(email);
+        if (res != null){
+            return res;
+        }else{
+            throw new UserNotFoundException(email);
+        }
     }
 
     @Override
-    public User createUser(User user) {
+    public User createUser(User user) throws TriddyServiceException {
         User res = userRepository.findByEmail(user.getEmail());
         if (res == null){
             user.setPassword(new Sha512Hash(user.getPassword()).toHex());
             return userRepository.save(user);
         }else{
-            return null;
+            throw new TriddyServiceException("User already exist");
+        }
+        
+    }
+
+    @Override
+    public void delUser(String user) throws UserNotFoundException {
+        User res = userRepository.findByEmail(user);
+        if (res != null){
+            userRepository.deleteByEmail(user);
+        }else{
+            throw new UserNotFoundException(user);
         }
     }
 
     @Override
-    public void delUser(String user) {
-        userRepository.deleteByEmail(user);
+    public void changeFullname(User user) throws UserNotFoundException {
+        User res = userRepository.findByEmail(user.getEmail());
+        if (res != null){
+            res.setFullname(user.getFullname());
+            userRepository.save(res);
+        }else{
+            throw new UserNotFoundException(user.getEmail());
+        }
+        
+        
     }
 
     @Override
-    public void changeFullname(User user) {
+    public void changeUniversity(User user) throws UserNotFoundException {
         User res = userRepository.findByEmail(user.getEmail());
-        res.setFullname(user.getFullname());
-        userRepository.save(res);
+        if (res != null){
+            res.setUniversity(user.getUniversity());
+            userRepository.save(res);  
+        }else{
+            throw new UserNotFoundException(user.getEmail());
+        }  
     }
 
     @Override
-    public void changeUniversity(User user) {
+    public void changeCareer(User user) throws UserNotFoundException {
         User res = userRepository.findByEmail(user.getEmail());
-        res.setUniversity(user.getUniversity());
-        userRepository.save(res);    
+        if (res != null){
+            res.setCareer(user.getCareer());
+            userRepository.save(res); 
+        }else{
+            throw new UserNotFoundException(user.getEmail());
+        }  
+           
     }
 
     @Override
-    public void changeCareer(User user) {
-        User res = userRepository.findByEmail(user.getEmail());
-        res.setCareer(user.getCareer());
-        userRepository.save(res);    
-    }
-
-    @Override
-    public void changePicture(User user) {
-        User res = userRepository.findByEmail(user.getEmail());
-        res.setPicture(user.getPicture());
-        userRepository.save(res);    
+    public void changePicture(User user) throws UserNotFoundException {
+        User res = userRepository.findByEmail(user.getEmail()); 
+        if (res != null){
+            res.setPicture(user.getPicture());
+            userRepository.save(res);  
+        }else{
+            throw new UserNotFoundException(user.getEmail());
+        }   
     }
     
 }
